+++ packages/e2e-tests/crossplatform.utils.ts
@@ -16,12 +16,12 @@ export class BuildSetup {
   }
 
   private getBinaryLocation() {
-    switch (process.env.TEST_SYSTEM) {
+    switch (process.platform) {
       case 'linux':
-        return `${__dirname}/Quiet/Quiet-0.19.17-alpha.0.AppImage`
-      case 'windows':
+        return `${__dirname}/Quiet/Quiet-1.0.0-alpha.13.AppImage`
+      case 'win32':
         return `${process.env.LOCALAPPDATA}\\Programs\\quiet\\Quiet.exe`
-      case 'mac':
+      case 'darwin':
         return '/Applications/Quiet.app/Contents/MacOS/Quiet'
       default:
         throw new Error('wrong SYSTEM env')
@@ -31,7 +31,8 @@ export class BuildSetup {
   public async createChromeDriver() {
     this.dataDir = (Math.random() * 10 ** 18).toString(36)
 
-    if (process.env.TEST_SYSTEM === 'windows') {
+    if (process.platform === 'win32') {
+      console.log('!WINDOWS!')
       exec(`cd %APPDATA% % & mkdir ${this.dataDir}`, e => console.log({ e }))
       this.child = spawn(
         `set DATA_DIR=${this.dataDir} & cd node_modules/.bin & chromedriver.cmd --port=${this.port}`,
@@ -42,15 +43,39 @@ export class BuildSetup {
       )
     } else {
       this.child = spawn(
-        `DATA_DIR=${this.dataDir} TEST_MODE=false node_modules/.bin/chromedriver --port=${this.port}`,
+        `DEBUG=backend DATA_DIR=${this.dataDir} node_modules/.bin/chromedriver --port=${this.port}`,
         [],
         {
-          shell: true
+          shell: true,
+          detached: false
         }
       )
     }
     // Extra time for chromedriver to setup
     await new Promise<void>(resolve => setTimeout(() => resolve(), 2000))
+
+    const killNine = () => {
+      exec(`kill -9 $(lsof -t -i:${this.port})`)
+      exec(`kill -9 $(lsof -t -i:${this.debugPort})`)
+    }
+
+    this.child.on('error', () => {
+      console.log('ERROR')
+      killNine()
+    })
+
+    this.child.on('exit', () => {
+      console.log('EXIT')
+      killNine()
+    })
+
+    this.child.on('close', () => {
+      console.log('CLOSE')
+      killNine()
+    })
+
+    this.child.on('message', data => console.log('message', data))
+    this.child.on('error', data => console.log('error', data))
 
     this.child.stdout.on('data', data => {
       console.log(`stdout:\n${data}`)
@@ -59,6 +84,26 @@ export class BuildSetup {
     this.child.stderr.on('data', data => {
       console.error(`stderr: ${data}`)
     })
+
+    this.child.stdin.on('data', data => {
+      console.error(`stdin: ${data}`)
+    })
+  }
+
+  public async getTorPid() {
+    const execAsync = async (cmd: string) => {
+      return await new Promise(resolve => {
+        exec(cmd, (error, stdout, stderr) => {
+          if (error) {
+            console.warn(error)
+          }
+          resolve(stdout || stderr)
+        })
+      })
+    }
+    const torPid = await execAsync('lsof -t -c tor')
+    console.log({ torPid })
+    return torPid
   }
 
   public getDriver() {
@@ -83,9 +128,10 @@ export class BuildSetup {
     return this.driver
   }
 
-  public killChromeDriver() {
+  public async killChromeDriver() {
     console.log('kill')
     this.child.kill()
+    await new Promise<void>(resolve => setTimeout(() => resolve(), 2000))
   }
 
   public async closeDriver() {