+++ packages/testcafe-browser-provider-electron/test/data/test-app-regular/client.js
@@ -1,3 +1,0 @@
-exports.foo = function () {
-    return 'HELLO ELECTRON!';
-};