+++ packages/desktop/CHANGELOG.md
@@ -3,6 +3,121 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0-alpha.15](https://github.com/TryQuiet/quiet/compare/quiet@1.0.0-alpha.14...quiet@1.0.0-alpha.15) (2023-02-16)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+# [1.0.0-alpha.14](https://github.com/TryQuiet/quiet/compare/quiet@1.0.0-alpha.13...quiet@1.0.0-alpha.14) (2023-02-15)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+# [1.0.0-alpha.13](https://github.com/TryQuiet/quiet/compare/quiet@1.0.0-alpha.12...quiet@1.0.0-alpha.13) (2023-02-14)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+# [1.0.0-alpha.12](https://github.com/TryQuiet/quiet/compare/quiet@1.0.0-alpha.11...quiet@1.0.0-alpha.12) (2023-02-13)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+# [1.0.0-alpha.11](https://github.com/TryQuiet/quiet/compare/quiet@1.0.0-alpha.10...quiet@1.0.0-alpha.11) (2023-02-13)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+# [1.0.0-alpha.10](https://github.com/TryQuiet/quiet/compare/quiet@1.0.0-alpha.9...quiet@1.0.0-alpha.10) (2023-02-13)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+# [1.0.0-alpha.9](https://github.com/TryQuiet/quiet/compare/quiet@1.0.0-alpha.8...quiet@1.0.0-alpha.9) (2023-02-13)
+
+
+### Reverts
+
+* Revert "Remove afterAllArtifactBuild for linux" ([975d0df](https://github.com/TryQuiet/quiet/commit/975d0df58494bdfba1270f6845152af4969e77ea))
+
+
+
+
+
+# [1.0.0-alpha.8](https://github.com/TryQuiet/quiet/compare/quiet@1.0.0-alpha.4...quiet@1.0.0-alpha.8) (2023-02-13)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+# [1.0.0-alpha.7](https://github.com/TryQuiet/quiet/compare/quiet@1.0.0-alpha.4...quiet@1.0.0-alpha.7) (2023-02-13)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+# [1.0.0-alpha.6](https://github.com/TryQuiet/quiet/compare/quiet@1.0.0-alpha.4...quiet@1.0.0-alpha.6) (2023-02-13)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+# [1.0.0-alpha.5](https://github.com/TryQuiet/quiet/compare/quiet@1.0.0-alpha.4...quiet@1.0.0-alpha.5) (2023-02-13)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+# [1.0.0-alpha.4](https://github.com/TryQuiet/quiet/compare/quiet@1.0.0-alpha.3...quiet@1.0.0-alpha.4) (2023-02-10)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+# [1.0.0-alpha.3](https://github.com/TryQuiet/quiet/compare/quiet@1.0.0-alpha.2...quiet@1.0.0-alpha.3) (2023-02-10)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
+# [1.0.0-alpha.2](https://github.com/TryQuiet/quiet/compare/quiet@1.0.0-alpha.1...quiet@1.0.0-alpha.2) (2023-02-09)
+
+**Note:** Version bump only for package quiet
+
+
+
+
+
 # [1.0.0-alpha.1](https://github.com/TryQuiet/quiet/compare/quiet@1.0.0-alpha.0...quiet@1.0.0-alpha.1) (2023-02-09)
 
 **Note:** Version bump only for package quiet